Add styles for new components
For the CSS naming convention was chosen the BEM methodology. It allows us to have a structured and scalable CSS styles. I used that a lot in the past and I do think that it fits here perfectly. If you want to learn more about it: https://en.bem.info/methodology/quick-start/.

Adds styles for the next components:
1) Heading 1;
2) Paragraph;
3) Button;
4) Form field;
5) Radio button;
6) Select for country codes.
CSS variables update was required for the component.

New JavaScript file was added for the country codes select. The select includes custom UI, and it's not achievable by using the standard `<select>` HTML tag because it's an operating system element, and we simply can't customize it. That's why the script was added. I was also thinking to do it with pure HTML + CSS, but I think it's not worth it. The logic becomes complicated and super hardcoded.

Also, image icons for some component were added.